+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Version 0.9.6
 
+- Added issue templates for GitHub
+
+- Added interactive plot for connectivity (iplotConn)
+
 - Added 'dynamicRates' option for NetStim populations
 
 - Added option to have a uniform distribution of rates for Vecstim pulses
@@ -8,7 +12,9 @@
 
 - Overhaul of contribution guide (CONTRIBUTING.md)
 
-- Improvements of documentation
+- Improved documentation
+
+- Fixed bugs and improved exception handling in plotConn 
 
 - Fixed loadSave V1 example model